pipeline test

+++ macos-install.sh
@@ -66,7 +66,7 @@ function install_fonts(){
     do
         res=$(null brew list --cask font-${font}-nerd-font)
         if [[ $? -ne 0 ]]; then
-            null sudo rm -f $HOME/Library/Fonts/${font}*
+            [[ -f $HOME/Library/Fonts/${font}* ]]] && null sudo rm -f $HOME/Library/Fonts/${font}*
             null brew tap homebrew/cask-fonts && 
             null brew install --cask font-${font}-nerd-font || return 1
         fi
